extract more stuff
