+++ src/main/java/dev/harrel/jsonschema/providers/JettisonNode.java
@@ -2,39 +2,21 @@ package dev.harrel.jsonschema.providers;
 
 import dev.harrel.jsonschema.JsonNode;
 import dev.harrel.jsonschema.JsonNodeFactory;
-import dev.harrel.jsonschema.SimpleType;
 import org.codehaus.jettison.json.JSONArray;
 import org.codehaus.jettison.json.JSONException;
 import org.codehaus.jettison.json.JSONObject;
 import org.codehaus.jettison.json.JSONTokener;
 
-import java.math.BigDecimal;
-import java.math.BigInteger;
 import java.util.*;
 
 public final class JettisonNode extends SimpleJsonNode {
-    private final Object node;
-    private final String jsonPointer;
-    private final SimpleType nodeType;
 
     private JettisonNode(Object node, String jsonPointer) {
-        this.node = Objects.requireNonNull(node);
-        this.jsonPointer = Objects.requireNonNull(jsonPointer);
-        this.nodeType = computeNodeType(node);
+        super(Objects.requireNonNull(node), jsonPointer);
     }
 
     public JettisonNode(Object node) {
         this(node, "");
-    }
-
-    @Override
-    public String getJsonPointer() {
-        return jsonPointer;
-    }
-
-    @Override
-    public SimpleType getNodeType() {
-        return nodeType;
     }
 
     @Override
@@ -45,24 +27,6 @@ public final class JettisonNode extends SimpleJsonNode {
     @Override
     public String asString() {
         return Objects.toString(isNull() ? null : node);
-    }
-
-    @Override
-    public BigInteger asInteger() {
-        if (node instanceof BigDecimal) {
-            return ((BigDecimal) node).toBigInteger();
-        } else {
-            return BigInteger.valueOf(((Number) node).longValue());
-        }
-    }
-
-    @Override
-    public BigDecimal asNumber() {
-        if (node instanceof BigDecimal) {
-            return (BigDecimal) node;
-        } else {
-            return BigDecimal.valueOf(((Number) node).longValue());
-        }
     }
 
     @Override
